% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konwersja_temp.R
\name{konwersja_temp}
\alias{konwersja_temp}
\title{Konwersja temperatur}
\usage{
konwersja_temp(temperatura_f)
}
\arguments{
\item{temperatura_f}{zawierajacy wartosci temperatury w stopniach F}
}
\value{
wektor numeryczny
}
\description{
Funkcja sluzaca do konwersji temperatury
ze stopni Fahrenheita do stopni Celsjusza
}
\examples{
konwersja_temp(75)
konwersja_temp(20)
}
