% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Potegowanie}
\usage{
pow(x, y)
}
\arguments{
\item{x}{Wartość numeryczna}

\item{y}{Wartość numeryczna (potęga)}
}
\value{
Wartość numeryczna
}
\description{
Wartosc x jest potegowana do potegi y.
}
\examples{

pow(2, 2)
pow(2, 3)

}
